refactor(shared): unify RefreshMode/PrOutput and IPC protocol DTOs in crate::shared
evaluation の CacheHint, daemon の RefreshMode, infrastructure の RefreshModeDto は
全て同じ意味の Hot/Warm/Terminal 三値で、互換のための From impl を 2 つ抱えていた。
PrOutput / PrOutputDto も同じく二重定義だった。

crate::shared モジュールを新設し、RefreshMode と protocol（PrOutput, EntryDto,
Request, Response）を集約。重複型と From 変換を全て削除し、context 間で同じ型を
直接共有する。

shared モジュールは DDD レイヤと独立し、scripts/check-layer-deps.sh の cross-context
正規表現に該当しないため script 変更不要。

Refs: #336
