+++ src/contexts/daemon/domain/cache.rs
@@ -1,9 +1,7 @@
 mod entry;
 mod port;
-mod refresh_mode;
 mod repo_id;
 
-pub use entry::{CacheEntry, PrOutput};
+pub use entry::CacheEntry;
 pub use port::CachePort;
-pub use refresh_mode::RefreshMode;
 pub use repo_id::RepoId;
